rename redirector to resolver
