+++ internal/services/resolver/resolver.go
@@ -1,4 +1,4 @@
-package redirector
+package resolver
 
 import (
 	"net/http"
@@ -27,9 +27,9 @@ func New(cfg *Config) *Service {
 	}
 }
 
-// Redirect reads URL path, retrieves corresponding URL from storage
+// Resolve reads URL path, retrieves corresponding URL from storage
 // and returns 307 response. It performs path validation before calling storage
-func (svc *Service) Redirect(w http.ResponseWriter, req *http.Request) {
+func (svc *Service) Resolve(w http.ResponseWriter, req *http.Request) {
 	var (
 		redirect string
 		err      error
